feat:implemented a match check for the player's index on the next turn
