+++ src/utils/nextPlayer.ts
@@ -1,0 +1,3 @@
+export const nextPlayer = {
+  index: 0
+}
